add writing to disk.

+++ src/inode.rs
@@ -368,7 +368,7 @@ impl Inode {
         let (block_index, offset_within_block) =
             get_inode_location(ext4, self.index)?;
         let block_size = ext4.0.superblock.block_size.to_u64();
-        let pos = u64::from(block_index) * block_size
+        let pos = block_index * block_size
             + u64::from(offset_within_block);
         self.update_inode_data(ext4);
         // Write only the data we've saved to avoid overwriting any unread info
